Split bot code from __main__.
